+++ slash/__init__.py
@@ -1,37 +1,16 @@
+# pylint: disable=unused-import
 from .__version__ import __version__
 from .cleanups import add_cleanup, add_critical_cleanup, add_success_only_cleanup, get_current_cleanup_phase, is_in_cleanup
 from .conf import config
 from .ctx import context
 from .ctx import g, session, test
+from .core.scope_manager import get_current_scope
 from .core.session import Session
 from .core.tagging import tag
 # assertions
 from . import assertions
 should = assertions
-from .assertions import (
-    allowing_exceptions,
-    assert_almost_equal,
-    assert_contains,
-    assert_equal,
-    assert_equals,
-    assert_false,
-    assert_in,
-    assert_is,
-    assert_is_none,
-    assert_empty,
-    assert_not_empty,
-    assert_is_not,
-    assert_is_not_none,
-    assert_isinstance,
-    assert_not_contain,
-    assert_not_contains,
-    assert_not_equal,
-    assert_not_equals,
-    assert_not_in,
-    assert_not_isinstance,
-    assert_raises,
-    assert_true,
-    )
+from .assertions import allowing_exceptions, assert_almost_equal, assert_raises
 from .core.test import Test
 from .core.test import abstract_test_class
 from .core.exclusions import exclude
@@ -41,7 +20,7 @@ from .core.fixtures.utils import fixture, use_fixtures, nofixtures, generator_fi
 from .core.requirements import requires
 from .utils import skip_test, skipped, add_error, add_failure, set_test_detail, repeat, register_skip_exception
 from .utils.interactive import start_interactive_shell
-from .warnings import ignore_warnings, clear_ignored_warnings
+from .warnings import ignore_warnings, ignored_warnings, clear_ignored_warnings
 from .runner import run_tests
 import logbook
 logger = logbook.Logger(__name__)