+++ slash/warnings.py
@@ -8,6 +8,7 @@ import warnings
 from . import hooks
 from .utils.warning_capture import warning_callback_context
 from .ctx import context
+from .exception_handling import handling_exceptions
 from contextlib import contextmanager
 
 
@@ -173,6 +174,20 @@ class _IgnoredWarning(object):
 
         return True
 
+    def __repr__(self):
+        extras = ""
+        if self.category:
+            extras += "category: {} ".format(self.category)
+        if self.message:
+            extras += "message: {} ".format(self.message)
+        if self.lineno:
+            extras += "lineno: {} ".format(self.lineno)
+        if self.filename:
+            extras += "filename: {} ".format(self.filename)
+        if not extras:
+            extras = "-"
+        return "<IgnoredWarning: {}>".format(extras.strip())
+
 _ignored_warnings = []
 
 
@@ -188,7 +203,21 @@ def ignore_warnings(category=None, message=None, filename=None, lineno=None):
 
     .. note:: Calling ignore_warnings() with no arguments will ignore **all** warnings
     """
-    _ignored_warnings.append(_IgnoredWarning(category=category, filename=filename, lineno=lineno, message=message))
+    iw = _IgnoredWarning(category=category, filename=filename, lineno=lineno, message=message)
+    _ignored_warnings.append(iw)
+    return iw
+
+
+@contextmanager
+def ignored_warnings(**kwargs):
+    iw = ignore_warnings(**kwargs)
+    with handling_exceptions():
+        yield
+    try:
+        _ignored_warnings.remove(iw)
+    except ValueError:
+        # In case clear_warnings was called inside this context, ValueError will be raised
+        pass
 
 
 def clear_ignored_warnings():